requirements.txt
numpy==1.21.5
Pillow==9.1.0
streamlit==1.8.1
opencv